lecture 3 completed, code formatter applied

+++ Assignment1/Card.h
@@ -9,11 +9,11 @@
 #import <Foundation/Foundation.h>
 
 @interface Card : NSObject
-@property (strong,nonatomic) NSString *contents;
+@property(strong, nonatomic) NSString *contents;
 
-@property (nonatomic, getter=isChosen) BOOL chosen;
-@property (nonatomic, getter=isMatched) BOOL matched;
+@property(nonatomic, getter=isChosen) BOOL chosen;
+@property(nonatomic, getter=isMatched) BOOL matched;
 
--(int) match:(NSArray*) otherCards;
+- (NSUInteger)match:(NSArray *)otherCards;
 
 @end
